migrator: review changes
* to squash
