+++ cds_ils/migrator/api.py
@@ -47,6 +47,68 @@ migrated_logger = logging.getLogger("migrated_records")
 records_logger = logging.getLogger("records_errored")
 
 
+def import_documents_from_record_file(sources, include):
+    """Import documents from records file generated by CDS-Migrator-Kit."""
+    include = include if include is None else include.split(",")
+    records = []
+    for idx, source in enumerate(sources, 1):
+        click.echo(
+            "({}/{}) Migrating documents in {}...".format(
+                idx, len(sources), source.name
+            )
+        )
+        model, provider = model_provider_by_rectype("document")
+        include_keys = None if include is None else include.split(",")
+        with click.progressbar(json.load(source).items()) as bar:
+            records = []
+            for key, parent in bar:
+                click.echo(
+                    'Importing document "{}"...'.format(parent["legacy_recid"])
+                )
+                if include_keys is None or key in include_keys:
+                    record = import_record(parent, model, provider)
+                    records.append(record)
+    # Index all new parent records
+    bulk_index_records(records)
+
+
+def import_documents_from_dump(sources, source_type, eager, include):
+    """Load records."""
+    include = include if include is None else include.split(",")
+    for idx, source in enumerate(sources, 1):
+        click.echo(
+            "({}/{}) Migrating documents in {}...".format(
+                idx, len(sources), source.name
+            )
+        )
+        data = json.load(source)
+        with click.progressbar(data) as records:
+            for item in records:
+                click.echo('Processing document "{}"...'.format(item["recid"]))
+                if include is None or str(item["recid"]) in include:
+                    try:
+                        import_document_from_dump(
+                            item, source_type, eager=eager
+                        )
+
+                        migrated_logger.warning(
+                            "#RECID {0}: OK".format(item["recid"])
+                        )
+                    except IlsValidationError as e:
+                        records_logger.error(
+                            "@RECID: {0} FATAL: {1}".format(
+                                item["recid"],
+                                str(e.original_exception.message),
+                            )
+                        )
+                    except Exception as e:
+                        records_logger.error(
+                            "@RECID: {0} ERROR: {1}".format(
+                                item["recid"], str(e)
+                            )
+                        )
+
+
 @contextmanager
 def commit():
     """Commit transaction or rollback in case of an exception."""
@@ -146,7 +208,9 @@ def import_multivolume(json_record):
         ]
     ]
 
-    if any(lists_lenghts) > len(volume_list):
+    too_many_volumes = any(lists_lenghts) > len(volume_list)
+
+    if too_many_volumes:
         raise ManualImportRequired(
             "Record has more additional volume information "
             "entries than the number of indicated volumes"
